fix: spearbit redstone issues

+++ test/adapter/redstone/RedstoneCoreOracleHelper.sol
@@ -88,7 +88,7 @@ contract RedstoneCoreOracleHelper is AdapterHelper {
         if (behaviors[Behavior.FeedReturnsZeroPrice]) {
             s.price = 0;
         } else if (behaviors[Behavior.FeedReturnsTooLargePrice]) {
-            s.price = bound(s.price, uint256(type(uint200).max) + 1, type(uint256).max);
+            s.price = bound(s.price, uint256(type(uint160).max) + 1, type(uint256).max);
         } else {
             s.price = bound(s.price, bounds.minPrice, bounds.maxPrice);
         }
